refactor(shim): PR-9.3 — 删 curation / datasets / thumb_cache 3 shim + 级联
batch 1 of 6 — dataset 域 3 个 top-level shim。

  删 studio/curation.py / datasets.py / thumb_cache.py (10 行 × 3 = 30 行)

caller 更新（13 处）：
  studio/api/responses.py             from .. import thumb_cache → ..services.dataset
  studio/api/routers/projects/
    curation.py                       from .... import curation, datasets, ... 拆出
    training.py                       同上 datasets
    ingestion.py                      同上 datasets
  studio/services/inference/upscaler.py from .. import thumb_cache → ..dataset
  studio/services/preprocess/duplicates.py from ... import curation 拆出
  studio/services/{preprocess/core,projects/projects,projects/versions,
    reg/{postprocess,builder,analysis}}.py
                                      6 个 from ...datasets import IMAGE_EXTS
                                      → from ...services.dataset.scan import IMAGE_EXTS
  studio/workers/{tag,reg_build}_worker.py from studio.datasets import → canonical
  studio/workers/preprocess_worker.py from studio import thumb_cache → canonical
  tests/test_curation.py              from studio import curation, ... 拆
  tests/test_datasets.py              from studio import datasets → canonical alias
  tests/test_upscaler.py              from studio import thumb_cache → canonical

回归：1592 passed（与 baseline 等价，2 pre-existing imagehash fail 不算）。
