+++ studio/api/routers/projects/curation.py
@@ -35,7 +35,8 @@ from ...schemas.curation import (
     RemoveRequest,
 )
 from ._shared import _publish_project_state
-from .... import curation, datasets, db, project_jobs, projects
+from .... import db, project_jobs, projects
+from ....services.dataset import curation, scan as datasets
 from ....event_bus import bus
 from ....services import duplicate_finder, preprocess_manifest
