+++ tests/test_curation.py
@@ -6,7 +6,9 @@ from pathlib import Path
 import pytest
 from PIL import Image, ImageDraw
 
-from studio import curation, db, projects, versions
+from studio import db
+from studio.services.dataset import curation
+from studio.services.projects import projects, versions
 from studio.services import duplicate_finder
 from studio.services import preprocess_manifest
